Fixed errors that prevented compilation

+++ src/azipe.hpp
@@ -25,7 +25,7 @@ namespace az{
     int aipe(const std::vector<cv::Mat_<float>>&,
                 const std::vector<cv::Mat_<float>>&,
                 cv::Mat_<float>&, float&, float&, float&,
-                int);
+                float);
 
     float limitYawRange(float);
